add performance tests back to check for loss during refactoring
Added the current state of the three tests I found in the VCS history, executed each test 10 times at the state of this commit. Added CSV file with the current results to compare later on. This commit can be purged from the history once we are done (and certain we did not lose any performance)

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>